Fix int cast in chanwg test
Should have been done in c1d1bd93d235cb78bc1422a183e40b9809d0175e.

+++ pkg/chanwg/waitgroup_test.go
@@ -71,7 +71,7 @@ func TestWaitGroupConcurrentDone(t *testing.T) {
 	t.Parallel()
 	var wg chanwg.WaitGroup
 	count := 100
-	wg.Add(int32(count))
+	wg.Add(count)
 
 	var doneCount int32
 	var mu sync.Mutex
